templates: wrapper: Template check_rust_call()
Currently, passing an error callback of type
std::function<std::exception(RustBuffer)> results in exceptions of type
std::exception always being thrown. This commit turns the function into
a template to allow concrete exception types to be thrown. Additionally,
the declaration is removed from the header, which _should_ be fine.